The base 11 encoding is now correct as far as I understand it.
... which may not be enough of course, but seems compatible with the reading
routine.

git-svn-id: http://svn.mkgmap.org.uk/mkgmap/trunk@1391 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/BufferedImgFileReader.java
@@ -226,7 +226,7 @@ public class BufferedImgFileReader implements ImgFileReader {
 
 		String digit = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 		while (num != 0) {
-			val.append(digit.charAt((int) (num % base)));
+			val.append(digit.charAt(num % base));
 			num /= base;
 		}
